feat(esl-utils): move groupBy into array utils
