+++ src/modules/esl-utils/misc/object.ts
@@ -134,14 +134,3 @@ export function deepMerge(...objects: any[]): any {
     return res;
   }, {});
 }
-
-/**
- * @returns object with a criteria value as a key and an array of original items that belongs to the current criteria value
- */
-export const groupBy = <T, V extends string | number>(array: T[], group: (item: T) => V): Record<V, T[]> => {
-  return array.reduce((obj: Record<V, T[]>, el: T) => {
-    const key = group(el);
-    obj[key] ? obj[key].push(el) : obj[key] = [el];
-    return obj;
-  }, {} as Record<V, T[]>);
-};
